Move the functions to get errno into their own file.
This is needed to populate the autodie::exception object.  I picked
the code from making $! to best simulate how the Perl code does it.

This function would probably be useful to normalize all the ways
errno is accessed in the code, possibly splitting it into a function
to get the number and string parts.