Allow caller information to be passed into autodie::exception->new
This is necessary when creating an autodie::exception object inside a
Perl op, autodie::exception cannot guess the caller correctly.  It's
probably faster and more accurate than guessing with caller.

Passing in undef is allowed and is not overwritten by a default.
caller will be undef if an IO function is called outside a
subroutine.

Also remove a vestigial use of errno.