Make autodie wrappers conditional on the version of Perl.
Rather than removing functions completely from autodie, which will
break backwards compatibility with older perls, autodie will know
which functions have been implemented internally and skip the wrapper.

It also sets $^H{"autodie/$func"} = 1 to signal internally to use
autodie for that function.  is_autodie_enabled() has been changed to
look for that.  This allows autodie to be turned on and off per
function even if that function is implemented internally.

For #13 and #10