Add exceptions to open() and close().
Caveats:

Implicit close on block exit does not throw an exception. #22

close() will warn in some cases, like passing it something that isn't
a filehandle. This is redundant and should be turned off with exceptions.

Sometimes open() will die in sv_2io() with a string. This will be
fixed by #20