Make Perl_vcroak() throw an exception.
All calls to Perl_croak() and friends will now throw an exception if
exceptions are on. This is a cheap way to get exceptions in a lot of
otherwise hard to reach places.

This is almost totally not tested. I tested substr() just because it's
easy to get it to throw an error at runtime.

The exception stringifies differently than without.  This may or may
not be a feature.