Eliminate the STATIC macro.
This eliminates a special thing people have to know about when coding
Perl.

It's not a compiler compatibility thing.  It was introduced in 76e3520
(in 1998) as part of PERL_OBJECT, I guess PERL_OBJECT wouldn't
work with static functions? PERL_OBJECT was removed in acfe0ab (in
2001) but STATIC lived on.

The macro is left in for compatibility with public XS code.

For #30